upgrade-script/dev-requirements.txt
bcrypt==4.1.3
build==1.2.1
cachecontrol==0.14.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
cryptography==43.0.0
distlib==0.3.8
dulwich==0.21.7
elasticsearch==7.13.0
exceptiongroup==1.2.2
fastjsonschema==2.20.0
filelock==3.15.4
idna==3.7
importlib-metadata==7.2.1
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
keyring==24.3.1
more-itertools==10.3.0
msgpack==1.0.8
numpy==1.26.4
packaging==24.1
pandas==2.2.0
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pyproject-hooks==1.1.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
rapidfuzz==3.9.3
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
tomli==2.0.1
tomlkit==0.12.5
trove-classifiers==2024.5.22
tzdata==2024.1
urllib3==1.26.18
virtualenv==20.26.3
zipp==3.19.2
